Add -dry flag, improve searches and logging.
This script often runs returning nor results for content in Sonarr or
Radarr. But running the commands back works. This is an indication of a
race condition where qBt is running the script before the *arr agents
have added them to the api/v3/queue/details endpoint.

This new version has some additional logging for debugging purposes and
checks sonarr and radar up to 30 times each searching for the id.